Make an early copy of the RSDP pulled out of the loader context
The code to try to expand the memory map in the MB2 tag structures
was causing he RSDP in the loader context to get corrupted. This
copies it out into a buffer the first time something gets the RSDP.
Note, SLBOOT is still broken. The original error was 0xc00020c1
which is invalid RSDP passed to the SINIT. Now the error is
0xc0000501 which is invalid RSDP checksum. Nothing is changing the
checksum (copious amounts of tracing show this) so it is unclear
what is wrong.

Signed-off-by: Ross Philipson <ross.philipson@oracle.com>